Implement initial finder + editor codepath
Does not do any syncing. Just executes the finder, gets the filepath,
and executes the editor.

This was actually surprisingly tough... mostly cause of how finicky fzf
is with stdin, stderr...

You need stderr given to fzf so it can render, since it uses stdout for
the chosen filename.

You need stdin for it to render the fuzzy-matched filepaths, believe it
or not... we had just assumed that was stdout or something and had tried
.stdout(Stdio::inherit()) to no avail... turns out it is stdin... I
guess that kind of makes sense when you realize stdin is technically
no different from stdout/stderr.